app/requirements.txt
streamlit==1.2.0
requests==2.26.0
pillow==8.3.1